pyproject.toml
[project]
name = "personal-site"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "flask>=3.0.3",
    "waitress>=3.0.0",
    "pyyaml>=6.0.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/personal_site"]

[tool.rye.scripts]
tw = "pnpm exec tailwindcss -i ./src/static/css/tailwind.css -o ./src/static/css/style.css --watch"
dev = "rye run python ./src/app.py"
server = "rye run python ./src/server.py"

